Optimize P_ChangeSwitchTexture (missing p_spec.h)

+++ FASTDOOM/p_spec.h
@@ -192,8 +192,7 @@ typedef struct
 
 extern button_t buttonlist[MAXBUTTONS];
 
-void P_ChangeSwitchTexture(line_t *line,
-                           int useAgain);
+void P_ChangeSwitchTexture(line_t *line, byte useAgain);
 
 void P_InitSwitchList(void);
